Set rp.description from an IExecutionListener so RP picks it up
Static class-load initializer ran too late — RP TestNG agent memoizes ListenerParameters on first event. Move the description-building logic into GenesisLaunchDescriptionListener (IExecutionListener) registered before ReportPortalTestNGListener in studio_central_projects_regression.xml, so rp.description is set before PropertiesLoader memoizes its overlay of System.getProperties().
