+++ src/test/resources/testng_suites/studio_central_projects_regression.xml
@@ -2,10 +2,11 @@
 <suite verbose="1" name="Central Projects" annotations="JDK" parallel="tests" thread-count="1">
 
     <listeners>
+        <listener class-name="configuration.listeners.GenesisLaunchDescriptionListener"/>
         <listener class-name="configuration.listeners.AnnotationTransformer"/>
         <listener class-name="com.epam.reportportal.testng.ReportPortalTestNGListener"/>
     </listeners>
 
-    <test name="TestLocalCentralProjects"><classes><class name="tests.ui.webstudio.client.TestLocalCentralProjects"/></classes></test>
+    <test name="TestLocalCentralProjects"><classes><class name="tests.ui.webstudio.client.TestGenesisProjectsApi"/></classes></test>
 
 </suite>
